reorganize device initialization for mem devs
