+++ fs/null.h
@@ -1,0 +1,9 @@
+#ifndef FS_NULL_H
+#define FS_NULL_H
+
+#include "kernel/fs.h"
+#include "fs/dev.h"
+
+extern struct dev_ops null_dev;
+
+#endif
