+++ meson.build
@@ -75,6 +75,8 @@ src = [
     'fs/pipe.c',
 
     'fs/dev.c',
+    'fs/mem.c',
+    'fs/null.c',
     'fs/tty.c',
     'fs/tty-real.c',
